fix MSVC: narrowing uint32_t->int for port, bump cutil-macros

+++ src/colibri.cpp
@@ -48,7 +48,7 @@ auto Colibri::connect(const jingle::Jingle& initiate_jingle, const bool secure) 
         .address   = uri_domain.data(),
         .path      = uri_path.data(),
         .protocol  = "xmpp",
-        .port      = ws_uri.port,
+        .port      = static_cast<int>(ws_uri.port),
         .ssl_level = secure ? ws::client::SSLLevel::Enable : ws::client::SSLLevel::TrustSelfSigned,
     }));
     // obj->ws_context.dump_packets = true;
